Add nav bar links, dropdown and refactor routing
This commit includes several modifications mainly focused on the navigation bar, as well as routing changes. Specifically, it adds various links to the nav bar, incorporates a dropdown for user login, and refactors the app routing. Furthermore, it implements a new landing page and updates other component CSS and HTML for compatibility and improved UX.

Signed-off-by: Geani Dumitrache <geani.g.dumitrache@gmail.com>