Add Swiper package and refactor landing page and footer
The Swiper package was added to the project dependencies to extend slider functionality. Simultaneously, significant refactoring was carried out on the landing page and the footer of the application. Changes included enhancing the visual design, introducing new sections, and improving responsiveness. Dashboard Navbar's functionality was also improved to include routing to the homepage.

Signed-off-by: Geani Dumitrache <geani.g.dumitrache@gmail.com>